refactor(providers): decouple BaseRecipeProvider from scraper singleton
Make BaseRecipeProvider pure — no IO, no recipeScraper import. Scraping
moved to RecipeImportOrchestrator (new) which accepts scrapeFromHtml as
injected function. Provider now only exposes extractImageFromHtml for
HTML-side image extraction. Subclasses (Quitoque, HelloFresh) inherit.
