+++ src/providers/ProviderRegistry.ts
@@ -74,9 +74,6 @@ export function findProviderForUrl(url: string): RecipeProvider | undefined {
 /**
  * Finds a provider by its unique identifier string
  *
- * Used by ImageRepair to retrieve the provider from a recipe's `sourceProvider`
- * field without needing a URL.
- *
  * @param providerId - The provider's unique ID (e.g., 'quitoque')
  * @returns The matching provider instance, or undefined if not found
  */
